+++ tasks.c
@@ -1205,7 +1205,7 @@ write_forked_task(forked_task ft)
 {
 	int             lineno = find_line_number(ft.program, ft.f_index, 0);
 
-	dbio_printf("0 %d %ld %d\n", lineno, ft.start_time, ft.id);
+	dbio_printf("0 %d %d %d\n", lineno, (int)ft.start_time, ft.id);
 	write_activ_as_pi(ft.a);
 	write_rt_env(ft.program->var_names, ft.rt_env, ft.program->num_var_names);
 	dbio_write_forked_program(ft.program, ft.f_index);
